update build scripts

+++ build.ps1
@@ -27,45 +27,54 @@ try {
 ###########################################################################
 # INSTALL .NET CORE CLI
 ###########################################################################
+$sdk = $args[0]
 
-Function Remove-PathVariable([string]$VariableToRemove)
+# only install sdk if -sdk was used
+if ($sdk -eq "-sdk")
 {
-    $path = [Environment]::GetEnvironmentVariable("PATH", "User")
-    if ($path -ne $null)
+    Write-Output("Using .NET SDK v" + $DotNetVersion)
+
+    Function Remove-PathVariable([string]$VariableToRemove)
     {
-        $newItems = $path.Split(';', [StringSplitOptions]::RemoveEmptyEntries) | Where-Object { "$($_)" -inotlike $VariableToRemove }
-        [Environment]::SetEnvironmentVariable("PATH", [System.String]::Join(';', $newItems), "User")
+        $path = [Environment]::GetEnvironmentVariable("PATH", "User")
+        if ($path -ne $null)
+        {
+            $newItems = $path.Split(';', [StringSplitOptions]::RemoveEmptyEntries) | Where-Object { "$($_)" -inotlike $VariableToRemove }
+            [Environment]::SetEnvironmentVariable("PATH", [System.String]::Join(';', $newItems), "User")
+        }
+
+        $path = [Environment]::GetEnvironmentVariable("PATH", "Process")
+        if ($path -ne $null)
+        {
+            $newItems = $path.Split(';', [StringSplitOptions]::RemoveEmptyEntries) | Where-Object { "$($_)" -inotlike $VariableToRemove }
+            [Environment]::SetEnvironmentVariable("PATH", [System.String]::Join(';', $newItems), "Process")
+        }
     }
 
-    $path = [Environment]::GetEnvironmentVariable("PATH", "Process")
-    if ($path -ne $null)
-    {
-        $newItems = $path.Split(';', [StringSplitOptions]::RemoveEmptyEntries) | Where-Object { "$($_)" -inotlike $VariableToRemove }
-        [Environment]::SetEnvironmentVariable("PATH", [System.String]::Join(';', $newItems), "Process")
+    # Get .NET Core CLI path if installed.
+    $FoundDotNetCliVersion = $null;
+    if (Get-Command dotnet -ErrorAction SilentlyContinue) {
+        $FoundDotNetCliVersion = dotnet --version;
     }
-}
 
-# Get .NET Core CLI path if installed.
-$FoundDotNetCliVersion = $null;
-if (Get-Command dotnet -ErrorAction SilentlyContinue) {
-    $FoundDotNetCliVersion = dotnet --version;
-}
+    if($FoundDotNetCliVersion -ne $DotNetVersion) {
+        $InstallPath = Join-Path $PSScriptRoot ".dotnet"
+        if (!(Test-Path $InstallPath)) {
+            mkdir -Force $InstallPath | Out-Null;
+        }
+        (New-Object System.Net.WebClient).DownloadFile($DotNetInstallerUri, "$InstallPath\dotnet-install.ps1");
+        & $InstallPath\dotnet-install.ps1 -Channel $DotNetChannel -Version $DotNetVersion -InstallDir $InstallPath;
 
-if($FoundDotNetCliVersion -ne $DotNetVersion) {
-    $InstallPath = Join-Path $PSScriptRoot ".dotnet"
-    if (!(Test-Path $InstallPath)) {
-        mkdir -Force $InstallPath | Out-Null;
+        Remove-PathVariable "$InstallPath"
+        $env:PATH = "$InstallPath;$env:PATH"
     }
-    (New-Object System.Net.WebClient).DownloadFile($DotNetInstallerUri, "$InstallPath\dotnet-install.ps1");
-    & $InstallPath\dotnet-install.ps1 -Channel $DotNetChannel -Version $DotNetVersion -InstallDir $InstallPath;
 
-    Remove-PathVariable "$InstallPath"
-    $env:PATH = "$InstallPath;$env:PATH"
+    $env:DOTNET_SKIP_FIRST_TIME_EXPERIENCE=1
+    $env:DOTNET_CLI_TELEMETRY_OPTOUT=1
 }
-
-$env:DOTNET_SKIP_FIRST_TIME_EXPERIENCE=1
-$env:DOTNET_CLI_TELEMETRY_OPTOUT=1
-
+else {
+    Write-Output("Using global .NET SDK")
+}
 
 ###########################################################################
 # INSTALL CAKE
